+++ lib/jdoc.rb
@@ -1,3 +1,4 @@
+require "active_support/core_ext/object/to_query"
 require "active_support/core_ext/string/inflections"
 require "cgi"
 require "erubis"